feat: moved proxy to separate package
